+++ src/main/java/org/mesdag/advjs/util/AdvJSEvents.java
@@ -10,7 +10,6 @@ import org.mesdag.advjs.advancement.AdvConfigureEventJS;
 import org.mesdag.advjs.advancement.AdvLockEventJS;
 import org.mesdag.advjs.integration.betteradvancements.BetterAdvEventJS;
 import org.mesdag.advjs.integration.betteradvancements.BetterAdvModifier;
-import org.mesdag.advjs.integration.revelationary.AdvRevelationEventJS;
 import org.mesdag.advjs.trigger.Trigger;
 import org.mesdag.advjs.trigger.custom.TriggerRegistryEventJS;
 
@@ -22,8 +21,6 @@ public interface AdvJSEvents {
     EventHandler ADVANCEMENT = GROUP.server("advancement", () -> AdvConfigureEventJS.class);
     EventHandler LOCK = GROUP.server("lock", () -> AdvLockEventJS.class);
     EventHandler TRIGGER = GROUP.startup("trigger", () -> TriggerRegistryEventJS.class);
-
-    EventHandler REVELATION = GROUP.client("revelation", () -> AdvRevelationEventJS.class);
     EventHandler BETTER_ADV = GROUP.server("betterAdv", () -> BetterAdvEventJS.class);
 
     static void postAdv(LootDataManager lootData) {